Make a public IO module
