+++ astropy/cosmology/core.py
@@ -16,7 +16,7 @@ from astropy.utils.decorators import classproperty
 from astropy.utils.metadata import MetaData
 
 from ._utils import all_parameters
-from .connect import (
+from .io._connect import (
     CosmologyFromFormat,
     CosmologyRead,
     CosmologyToFormat,
@@ -204,7 +204,7 @@ class Cosmology(metaclass=ABCMeta):
         _COSMOLOGY_CLASSES[cls.__qualname__] = cls
 
         # register to YAML
-        from astropy.cosmology._io.yaml import register_cosmology_yaml
+        from astropy.cosmology.io._builtin.yaml import register_cosmology_yaml
 
         register_cosmology_yaml(cls)
