+++ astropy/cosmology/funcs/tests/test_comparison.py
@@ -8,8 +8,6 @@ import numpy as np
 import pytest
 
 from astropy.cosmology import Cosmology, FlatCosmologyMixin, Planck18, cosmology_equal
-from astropy.cosmology._io.tests.base import ToFromTestMixinBase
-from astropy.cosmology.connect import convert_registry
 from astropy.cosmology.funcs.comparison import (
     _CANT_BROADCAST,
     _cosmology_not_equal,
@@ -17,13 +15,15 @@ from astropy.cosmology.funcs.comparison import (
     _parse_format,
     _parse_formats,
 )
+from astropy.cosmology.io import convert_registry
+from astropy.cosmology.io.tests.base import ToFromTestMixinBase
 
 
 class ComparisonFunctionTestBase(ToFromTestMixinBase):
     """Tests for cosmology comparison functions.
 
     This class inherits from
-    `astropy.cosmology._io.tests.base.ToFromTestMixinBase` because the cosmology
+    `astropy.cosmology.io.tests.base.ToFromTestMixinBase` because the cosmology
     comparison functions all have a kwarg ``format`` that allow the arguments to
     be converted to a |Cosmology| using the ``to_format`` architecture.
