+++ astropy/cosmology/io/_builtin/ecsv.py
@@ -161,8 +161,8 @@ Additional keyword arguments are passed to ``QTable.read`` and ``QTable.write``.
 
 import astropy.cosmology.units as cu
 import astropy.units as u
-from astropy.cosmology.connect import readwrite_registry
 from astropy.cosmology.core import Cosmology
+from astropy.cosmology.io._connect import readwrite_registry
 from astropy.table import QTable
 
 from .table import from_table, to_table
