+++ astropy/cosmology/io/_builtin/html.py
@@ -87,8 +87,8 @@ enable this, set ``latex_names=True``.
 
 import astropy.cosmology.units as cu
 import astropy.units as u
-from astropy.cosmology.connect import readwrite_registry
 from astropy.cosmology.core import Cosmology
+from astropy.cosmology.io._connect import readwrite_registry
 from astropy.cosmology.parameter import Parameter
 from astropy.table import QTable
