+++ astropy/cosmology/io/_builtin/model.py
@@ -34,8 +34,8 @@ from dataclasses import replace
 
 import numpy as np
 
-from astropy.cosmology.connect import convert_registry
 from astropy.cosmology.core import Cosmology
+from astropy.cosmology.io._connect import convert_registry
 from astropy.modeling import FittableModel, Model
 from astropy.utils.decorators import classproperty
