+++ astropy/cosmology/io/_builtin/yaml.py
@@ -19,8 +19,8 @@ require YAML serialization.
 
 import astropy.cosmology.units as cu
 import astropy.units as u
-from astropy.cosmology.connect import convert_registry
 from astropy.cosmology.core import _COSMOLOGY_CLASSES, Cosmology
+from astropy.cosmology.io._connect import convert_registry
 from astropy.io.misc.yaml import AstropyDumper, AstropyLoader, dump, load
 
 from .mapping import from_mapping
