+++ astropy/cosmology/io/tests/test_cosmology.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from astropy.cosmology._io.cosmology import from_cosmology, to_cosmology
+from astropy.cosmology.io._builtin.cosmology import from_cosmology, to_cosmology
 
 from .base import IODirectTestBase, ToFromTestMixinBase
