+++ astropy/cosmology/io/tests/test_ecsv.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from astropy.cosmology._io.ecsv import read_ecsv, write_ecsv
 from astropy.cosmology.core import _COSMOLOGY_CLASSES
+from astropy.cosmology.io._builtin.ecsv import read_ecsv, write_ecsv
 from astropy.table import QTable, Table, vstack
 
 from .base import ReadWriteDirectTestBase, ReadWriteTestMixinBase
