+++ astropy/cosmology/io/tests/test_json.py
@@ -8,8 +8,8 @@ import pytest
 
 import astropy.units as u
 from astropy.cosmology import units as cu
-from astropy.cosmology.connect import readwrite_registry
 from astropy.cosmology.core import Cosmology
+from astropy.cosmology.io._connect import readwrite_registry
 
 from .base import ReadWriteDirectTestBase, ReadWriteTestMixinBase
