+++ astropy/cosmology/io/tests/test_mapping.py
@@ -6,7 +6,7 @@ import numpy as np
 import pytest
 
 from astropy.cosmology import Cosmology
-from astropy.cosmology._io.mapping import from_mapping, to_mapping
+from astropy.cosmology.io._builtin.mapping import from_mapping, to_mapping
 
 from .base import ToFromDirectTestBase, ToFromTestMixinBase
