+++ astropy/cosmology/io/tests/test_row.py
@@ -2,8 +2,8 @@
 
 import pytest
 
-from astropy.cosmology._io.row import from_row, to_row
 from astropy.cosmology.core import _COSMOLOGY_CLASSES, Cosmology
+from astropy.cosmology.io._builtin.row import from_row, to_row
 from astropy.table import Row
 
 from .base import ToFromDirectTestBase, ToFromTestMixinBase
